+++ theories/fset.v
@@ -764,6 +764,11 @@ End Properties.
 
 Arguments fsubsetP {_ _ _}.
 Arguments fset2P {_ _ _}.
+Arguments fsetIidPl {T s1 s2}.
+Arguments fsetIidPr {T s1 s2}.
+Arguments fsetUidPl {T s1 s2}.
+Arguments fsetUidPr {T s1 s2}.
+Arguments fsetDidPl {T s1 s2}.
 
 Section setOpsAlgebra.