+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 - Type of finitely supported functions `ffun`.
 
+## Changed
+
+- Implicit arguments for `fsetIidPl`, `fsetIidPr`, `fsetUidPl`, `fsetUidPr`,
+  `fsetDidPl`.
+
 # 0.2.2 (2020/08/13)
 
 - Fix compatibility issues with Coq 8.12 and Ssreflect 1.11.